refactor: extract HslColorPicker from color_picker_row
